requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
fpdf==1.7.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
Pillow==9.4.0
protobuf==3.20.3
pymongo==4.3.3
requests==2.28.2
six==1.16.0
soupsieve==2.4
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.13.0
